requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
numpy
pillow
torch
torchvision
torchinfo
tensorboard